Fixed auxiliary/scanner/http/frontpage_login error.
This is a fix of Bug #6354 of the Metasploit Issue Tracker. There
was an error when trying to write results in the database with an
empty port number.

The code has been modified to write a value if port is empty and
to write the port value without the colon at the beginning if it is
not empty.

Reported by: Nelson LeBlanc
Signed off by: Silviu Popescu <silviupopescu1990@gmail.com>