Get fascist about 50/72 git commit messages
Git commits must always, always, always stick to 50 characters or less
for formatting. This is the default convention followed by many existing
git projects, and many git tools assume these widths for display
formatting.

If new commits don't adhere to this, they should be rejected out of
hand with a brief explination and link to http://r-7.co/OGM3OK . It's
usually not that hard for the committer to either git commit --amend or
to merge commits over to a new branch with a new commit message.