Allow use of alternate credentials
This commit modifies the initial WMI execution wrapper to allow
for the use of alternate credentials.

TODO: resolve local execution issue and adjust wrapper. Calls to
remote machines with admin privileges result in a UAC bypassed
state. If the same can be accomplished locally, then this method
could be used to perform a UAC bypass without any binaries.