Cleanup send_request_cgi hackery for mitm proxy
This commit replaces the Msf::Exploit::HttpClient overload mess
with proper methods for proxy_request_cgi and proxy_connect to be
used within the module's runtime. These methods perform the
request_cgi method with options from the modified headers instead
of using datastore values to instantiate the connection and params.