Add an NTP enumeration module
This module gets the current time and unique ID of the NTP server.
Requires 'net/ntp' and provides overloads for running over Rex
sockets and displaying the raw payload data from the response.
There may be more to query than just the ID, but it does appear to
be a unique identifier which can be used for identifying common
hosts with multiple interfaces.

Setting the CHOST variable will allow scanner threads to re-use
the bound socket provided it can reach the target host.
User can also set the CLIENT_TIME variable to spoof request time.