Add a CONTRIBUTING file. Delete HACKING.
The CONTRIBUTING.md functions as a replacement for HACKING. Having a
CONTRIBUTING file will enable Github to show a prompt with a
link to the file to any user who opens a pull request. This helps users
check that their changes comply with the contribution guidelines and
information found on the wiki. In addition CONTRIBUTING contains links
from the Metasploit Environment guide wiki page to expand on its own
contents.

*NOTE*: Links to HACKING will need to be fixed on the wiki before this goes into
trunk.

Github blog post about contributing guidelines:
https://github.com/blog/1184-contributing-guidelines