New meterpreter extension called ocioralog.
With the ocioralog meterpreter extension it is possible to hook
critical functions of the OCI driver. Currently the following functions
are supported:
- OCIAttrSet - Called when a user connects to a database and takes the
username and the cleartext password as parameters (when the user logs
in).
- OCIServerAttach - Called when the connection string is passed to the
TNS listener (when the user logs in).
- OCIStmtExecute - Called every time when an SQL statement is executed
from the client.

Hijacking these 3 functions makes it possible to dump the following
information:
- Username and password in time of login
- Oracle connection strings in time of login
- Username and password in time of every SQL statement execution!