Adds Notes to Creds
This adds notes to Creds.  This is useful
for recording meta information about an account
such as groups an account belongs to or other things
from post intel gathering.  The migration file in
data/sql should work but was unaware of the proper method
to trigger a migration.

This patch allows for seemless integration when
specifying :user and :pass when calling report_note