SAP RFC aux and exploit modules
The majority of the modules rely on the SAP NW RFC SDK and require the
Ruby wrapper ‘nwrfc’ by Martin Ceronio. Unfortunately the NW RFC SDK is
available only to those who have access to the SAP Service Marketplace
(SMP). In order to get access to the SAP Marketplace you need an S-ID,
password and customer number. Alternatively the required library files
(such as libsapnwrfc) can be extracted from a SAP system (such as the
freely available test drive systems).

These third party requirements are one of the reasons that the
submission to the Metasploit Framework proved problematic.

The modules are based on, inspired by, or are a port of plugins
available in the Onapsis Bizploit Opensource ERP Penetration Testing
framework - http://www.onapsis.com/research-free-solutions.php.

See
http://labs.mwrinfosecurity.com/blog/2012/04/27/mwr-sap-metasploit-modul
es for more info.