Made the Outlook module work with Outlook Express
I made changes to this module to add support for Outlook 98, Outlook 2000, and Outlook Express. It's just a different key to search in the registry. The structure is the same, with the exception that Outlook Express doesn't have a "Display Name", so I use "Account Name" instead.

Express also doesn't have a NextAccountID field, so I changed the logic slightly such that it works for all versions of Outlook. (by retrieving accounts and checking for nil)