cleaned up and added lots of data.nil/data.empty? checks
egypt, you really need to review this, or someone else.  I'm not sure
they're all needed.  I didn't see good places to add checks in the
parser files themselves, most of the changes to parser files in this
commit are cosmetic.

The meat of the commit is in db.rb and that needs a good review.