Update sudo module
This commit allows passing a password to the sudo command as
RM7581 does, and also attempts to pull credentials for the current
user from the workspace credstore if available.

If credentials are available, sudo is best run with a TTY shell
which can be obtained with python or expect (example can be seen
in post_multi_getty branch).