Import correct password type in Metasploit DB.
The Metasploit XML file parser sets the :ptype option on
cred_data because that is the name of the XML tag. The
report_auth_info method is expecting an option of :type.
Since the :type option is not given, all imported passwords
are of type "password". I modified the XML parser to set the
:type option instead of the :ptype option.