Return Msf::RPC::Client and Service in Rex::Proto
Msf::RPC included several classes of use to Rex and Msf for
interacting with XMLRPC. The Client and Service classes are
useful for API interactions, servers, and fuzzers. All of the code
pertaining to the XMLRPC libraries was purged with msgpack coming
in, mostly at 4829968107d1015436e9ac7357be69df7158c4be.

This commit returns the client.rb and service.rb files, places them
in rex/proto/xmlrpc with the appropriate xmlrpc.rb require hooks.