Keepalive and host resolution updates
Testing shows that reverse proxies do not like having the host IP
passed to them. Since we want to keep httpclient flexible and able
to hit IPs which may not actually resolve to the hostname/vhost
the request headers are rewritten again prior to the request being
made over the ip-based established connection.

Additionally, a check has been added when making a request to check
if the session to be reused is actually connected to the same host
and vhost. If not, a new httpclient is created for the request.

Still outstanding is communication with some JS user interfaces
which want to do asynchronous transfer. There are stubs for running
an additional monitor thread mechanism to watch client connections
for data, but this does not seem to work. Current test target is
the MikroTik webfig UI which passes binary data across the session,
all the while marking it as plain/text.