safari_metadata_archive.rb - Use Dir.chdir()
Use Dir.chdir() as supplied by feedback from limhoff-r7

However, the best way to fix this would be to not touch the current
working directory at all, as it's not thread safe (especially in
the case of msfrpcd). I don't have a vulnerable machine to test
that type of change, however.