Fix undefined method error for HTTP/HTTPS handler
This commit fixes the following error that may occur when an HTTP or
HTTPS handler is setup and a client connects to the web server to
request the root directory.

The error is:

Proc::on_request: NoMethodError: undefined method `%' for nil:NilClass

The error occurs when Rex::Text.checksum8(uri_strip) is called with an
empty string. This commit adds a check for an empty string and forces
the function to return the original requested URI.