Check whether systemsetup is actually runnable
On my system (10.8.4) I pass the admin check, but still fail to run systemsetup.
$ /usr/sbin/systemsetup
You need administrator access to run this tool... exiting!

Added a check for this.

Also made user_in_admin_group in "check" return Exploit::CheckCode::Safe, which it didn't do before and I assume it was supposed to.