Remove fastlib support
[#56005352]

HD agreed that I don't need to support fastlib for module loading and
the new module caching for the time being, so remove fastlib support so
people don't get the mistaken idea that it is supported.  Fastlib
support will be resurrect at such a time that HD or someone of he
designates has time to do HD's vision for Fastlib justice.  Fastlib will
live on in unstable for the time being.