Minor bug fix within generated cleanup resource file.
The windows pathname needs to have the '\' separators within the path
escaped.  Also clarified the return status from the "target_exec"
function; returning true/nil seems potentially counter intuitive so
changed to true/false.  The comment was also incorrect (no attempt was
made to return the pid), and the invoking code was checking for the
'nil', which whilst safe re-enforced the belief that a non-nil PID
value might be expected upon success.