Small improvements
Include ARCH_X86_64 in spec

Default Exploit::Exe arch should be dependent upon the payload.arch
not the TARGET arch because we can set the payload to whatever we
want but we must always have the correct shellcode in the correct
binary at the end of the day.

Do we want to print a warning in Exploits when the payload.arch
doesn't match the TARGET arch?