Single Payload Module::Class
[#56005352]

For Msf::Payloads with a single module ancestor that is a single
payload, generate the same Msf::Payload subclass as Msf::PayloadSet used
to and store the metadata to and Mdm::Module::Class in the database.
The beginnings of the process and naming for staged payload classes is
also in this commit as it helped find the common code between
non-payloads, single payloads and staged payloads.