travis-ci line-limit fix
[#58198114]

Travis-CI has a 10000 line limit which metasploit-framework was easily
busting when using the documentation format as it has over 10000
examples and each example takes more than one line (due to contexts,
etc) on average.  Switch to using the Fuubar formatter, which will use a
progress bar when there are no error, but print failures immediately,
and then continue with the progress bar.  Pending examples are printed
at the end.